Quick Refactor
